types fixes & test
