+++ src/components/Inputs/Select/Select.tsx
@@ -4,10 +4,9 @@ import React, { useRef, useState } from 'react';
 import { useOnOutsideClick } from '../../../hooks';
 import { ChevronDown, ChevronUp } from '../../../icons/regular';
 
-import { GeneralProps } from '../types';
-import { ColorsMap, SelectOption, SelectSpecificProps } from './SelectTypes';
+import { ColorsMap, GeneralProps, SelectOption, SelectSpecificProps } from '../types';
 
-type SelectProps = Omit<GeneralProps, 'value'> & SelectSpecificProps;
+type SelectProps = Omit<GeneralProps, 'value'> & Omit<GeneralProps, 'onChange'>& SelectSpecificProps;
 
 const Select = ({ searchable = false, options, defaultValue, size = 'large', state = 'idle', placeholder = '', onChange }:SelectProps) => {
   const [selectedOption, setSelectedOption] = useState(defaultValue);
