+++ src/components/Inputs/types.ts
@@ -10,6 +10,17 @@ export interface GeneralProps {
   onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
 }
 
+export interface SelectOption{
+  value:string;
+  label:string
+}
+export interface SelectSpecificProps {
+  options: SelectOption[];
+  defaultValue?:SelectOption;
+  searchable?:boolean;
+  // eslint-disable-next-line no-unused-vars
+  onChange: (option: SelectOption) => void;
+}
 export interface ColorsMap {
   [key: string]: string;
 }
